Making annotations PSR-0 autoloadable
